refactor(knowledge): rename pre-processors to post-processors
The parsing module was named pre-processors.ts but its contents are
post-processors (registry, applyPostProcessors, getAllPostProcessors)
that run after import. Align the file name with what it does and update
the two imports.
